refactor:extract GameTable to component
