+++ src/components/index.ts
@@ -2,3 +2,4 @@ export * from './Footer';
 export * from './Header';
 export * from './GameHeader';
 export * from './Timer';
+export * from './GameTable';
